% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewBradleyTerryScalable.R
\docType{package}
\name{NewBradleyTerryScalable}
\alias{NewBradleyTerryScalable}
\title{A package for fitting the Bradley-Terry model to (potentially) large and sparse data sets.}
\description{
A package for fitting the Bradley-Terry model to (potentially) large and sparse data sets.
}
